Support multiline customHeader
Support for passing array of strings as `customHeader` parameter. May be very useful when customHeader is used to define multiple `type`s, which are later used as return types in `returnTypes`. It it possible to use define types in a single line, dividing them with `\n`, but the new approach is way more readable, i.e.:
```
{
  "tspegjs": {
    "customHeader": [
      "type VehicleType = {type: string, model: string, registration: string}",
      "type UfoType = {type: string, diameter: number}"
    ]
  },
  "returnTypes": {
    "CAR": "VehicleType",
    "MOTORCYCLE": "VehicleType",
    "UFO": "UfoType"
  }
}
```